+++ Category/README.md
@@ -23,8 +23,12 @@
   - [imos](./imos) いもす法。累積和を応用した加算テクニック
   - [shakutori](./imos) しゃくとり法配列の添字2つを交互に動かして効率的に走査する手法
   - [division scheduling](./imos) 配列の特定区間。貪欲法を使う
-- DP 動的計画法
-  - [knapsack](./knapsack) ナップサック問題
+- DP 動的計画法(小さな問題の解決をメモして大きな問題を効率的にとく)
+  - 最大最小: その範囲での最大最小
+  - yes/no: その条件下でのtrue/false
+  - 組み合わせ: その条件下での組み合わせ数
+  - [chokudai](./dp/chokudai) 文字列中のキーワード組み合わせパターン数検索
+  - [70_100to105](../Atcoder/BootCamp/Easy/70_100to105) ナップサック問題
 
 
 ## Requirements
@@ -42,5 +46,8 @@
 
 [Atcoder problems](https://kenkoooo.com/atcoder/#/table/)
 [AtCoderの問題を分類しました【カテゴリ】 zenn](https://zenn.dev/koyanagihitoshi/books/atcoder-classification-6)
-
+問題をカテゴリ分類してくれている
+[EDPC Educational DP Contest](https://atcoder.jp/contests/dp/tasks)   
+DP問題ばかりのコンテスト
 [競技プログラミングにおける動的計画法問題まとめ はまやん](https://blog.hamayanhamayan.com/entry/2017/02/27/021246)
+DPに関する考察まとめ
